Use isAuth middleware

+++ app/item/router.js
@@ -11,6 +11,8 @@ router.get("/", async (_, res) => {
 
 router.post("/", async (req, res) => {
   try {
+    if (!req.isAuth) throw new Error("Not authenticated");
+
     const newItem = req.body;
     const id = await controller.create(newItem);
 
@@ -22,6 +24,8 @@ router.post("/", async (req, res) => {
 
 router.put("/:id", async (req, res) => {
   try {
+    if (!req.isAuth) throw new Error("Not authenticated");
+
     const { id } = req.params;
     const updatedItem = req.body;
     await controller.update(id, updatedItem);
@@ -34,6 +38,8 @@ router.put("/:id", async (req, res) => {
 
 router.delete("/:id", async (req, res) => {
   try {
+    if (!req.isAuth) throw new Error("Not authenticated");
+
     const { id } = req.params;
     await controller.delete(id);
   } catch (error) {
